test(slider): extract pointer tests into separate test files
- Extract pointer tests from slider.test.ts to slider-pointer.test.ts
- Extract pointer tests from range-slider.test.ts to range-slider-pointer.test.ts
- Organize tests into nested suites: value, change event, focus, disabled, readonly
- Move change event spy to beforeEach blocks
